+++ CHANGELOG.md
@@ -66,6 +66,8 @@ All notable changes to this project will be documented here. The format is based
 - There is now an (extremely EXPERIMENTAL) option to visualize plane contact forces in the 3D visualizer (#449)
 - "Toggle Frame Visibility" (model-wide) now appears as an option when right-clicking joints, so users can more
   easily know it's an option (#50)
+- The model editor's 3D viewport now has the same keybinds as the mesh importer (e.g. press G to grab, R to rotate,
+  left-arrow to orbit the camera, etc.; #44)
 
 ## [0.3.2] - 2023/01/09